+++ src/Eloquent/Relations/Concerns/IsAranguentRelation.php
@@ -4,21 +4,22 @@ namespace LaravelFreelancerNL\Aranguent\Eloquent\Relations\Concerns;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Query\Expression;
-use LaravelFreelancerNL\FluentAQL\QueryBuilder;
 
 trait IsAranguentRelation
 {
     /**
      * Add the constraints for a relationship count query.
      *
-     * @param  \Illuminate\Database\Eloquent\Builder  $query
-     * @param  \Illuminate\Database\Eloquent\Builder  $parentQuery
-     * @return \Illuminate\Database\Eloquent\Builder
+     * @param  Builder  $query
+     * @param  Builder  $parentQuery
+     * @return Builder
      */
     public function getRelationExistenceCountQuery(Builder $query, Builder $parentQuery)
     {
         return $this->getRelationExistenceQuery(
-            $query, $parentQuery, new Expression('*')
+            $query,
+            $parentQuery,
+            new Expression('*')
         );
     }
 }
