+++ src/Query/Concerns/HasAliases.php
@@ -22,6 +22,9 @@ trait HasAliases
      */
     public function registerTableAlias(string $table, string $alias = null): string
     {
+        if ($alias == null && stripos($table, ' as ') !== false) {
+            [$table, $alias] = explode(' as ', $table);
+        }
         if ($alias == null) {
             $alias = $this->generateTableAlias($table);
         }
@@ -113,7 +116,7 @@ trait HasAliases
      * @param  string|null  $alias
      * @return bool
      */
-    public function registerColumnAlias(string $column, string $alias = null) : bool
+    public function registerColumnAlias(string $column, string $alias = null): bool
     {
         if (preg_match("/\sas\s/i", $column)) {
             [$column, $alias] = $this->extractAlias($column);
@@ -152,7 +155,7 @@ trait HasAliases
             $table = $query->from;
         }
 
-        if ((is_string($column) || is_numeric($column)) && key_exists($column, $query->variables) ) {
+        if ((is_string($column) || is_numeric($column)) && key_exists($column, $query->variables)) {
             return $column;
         }
 
@@ -167,16 +170,28 @@ trait HasAliases
 
         //We check for an existing alias to determine of the first reference is a table.
         // In which case we replace it with the alias.
+        $references = $this->normalizeColumnReferences($references, $table);
+
+        return implode('.', $references);
+    }
+
+    /**
+     * @param $references
+     * @param  null  $table
+     * @return mixed
+     */
+    protected function normalizeColumnReferences($references, $table = null)
+    {
         $tableAlias = $this->getTableAlias($references[0]);
         if (isset($tableAlias)) {
             $references[0] = $tableAlias;
         }
 
-        if ($tableAlias === null && ! $this->isTableAlias($references[0])) {
+        if ($tableAlias === null && $table != null && ! $this->isTableAlias($references[0])) {
             $tableAlias = $this->generateTableAlias($table);
             array_unshift($references, $tableAlias);
         }
 
-        return implode('.', $references);
+        return $references;
     }
 }
