+++ src/Query/Grammar.php
@@ -5,6 +5,7 @@ namespace LaravelFreelancerNL\Aranguent\Query;
 use Illuminate\Support\Arr;
 use Illuminate\Support\Traits\Macroable;
 use LaravelFreelancerNL\Aranguent\Query\Concerns\CompilesAggregates;
+use LaravelFreelancerNL\Aranguent\Query\Concerns\CompilesColumns;
 use LaravelFreelancerNL\Aranguent\Query\Concerns\CompilesJoins;
 use LaravelFreelancerNL\Aranguent\Query\Concerns\CompilesWhereClauses;
 use LaravelFreelancerNL\Aranguent\Query\Concerns\HasAliases;
@@ -19,6 +20,7 @@ use LaravelFreelancerNL\FluentAQL\Grammar as FluentAqlGrammar;
 class Grammar extends FluentAqlGrammar
 {
     use CompilesAggregates;
+    use CompilesColumns;
     use CompilesJoins;
     use CompilesWhereClauses;
     use HasAliases;
@@ -312,90 +314,6 @@ class Grammar extends FluentAqlGrammar
         return $builder;
     }
 
-    /**
-     * Compile the "RETURN" portion of the query.
-     *
-     * @param Builder $builder
-     * @param array   $columns
-     *
-     * @return Builder
-     */
-    protected function compileColumns(Builder $builder, array $columns): Builder
-    {
-        $returnDocs = [];
-        $returnAttributes = [];
-        $values = [];
-
-        // Prepare columns
-        foreach ($columns as $column) {
-            // Extract rows
-            if (substr($column, strlen($column) - 2)  === '.*') {
-                $table = substr($column, 0, strlen($column) - 2);
-                $returnDocs[] = $this->getTableAlias($table);
-
-                continue;
-            }
-
-            if ($column != null && $column != '*') {
-                [$column, $alias] = $this->extractAlias($column);
-
-                $returnAttributes[$alias] = $this->normalizeColumn($builder, $column);
-            }
-        }
-        $values = $this->determineReturnValues($builder, $returnAttributes, $returnDocs);
-        $builder->aqb = $builder->aqb->return($values, (bool) $builder->distinct);
-
-        return $builder;
-    }
-
-    protected function determineReturnValues($builder, $returnAttributes = [], $returnDocs = [])
-    {
-        $values = [];
-
-        if (! empty($returnAttributes)) {
-            $values = $returnAttributes;
-        }
-
-        // If there is just one attribute/column given we assume that you want a list of values
-        //  instead of a list of objects
-        if (count($returnAttributes) == 1 && empty($returnDocs)) {
-            $values = reset($returnAttributes);
-        }
-
-        if (! empty($returnAttributes) && ! empty($returnDocs)) {
-            $returnDocs[] = $returnAttributes;
-        }
-
-        if (! empty($returnDocs)) {
-            $values = $builder->aqb->merge(...$returnDocs);
-        }
-
-        if ($builder->aggregate !== null) {
-            $values = ['aggregate' => 'aggregateResult'];
-        }
-
-        if (empty($values)) {
-            $values = $this->getTableAlias($builder->from);
-            if (is_array($builder->joins) && !empty($builder->joins)) {
-                $values = $this->mergeJoinResults($builder, $values);
-            }
-        }
-
-        return $values;
-    }
-
-
-    protected function mergeJoinResults($builder, $baseTable)
-    {
-        $tablesToJoin = [];
-        foreach ($builder->joins as $join) {
-            $tablesToJoin[] = $this->getTableAlias($join->table);
-        }
-        $tablesToJoin = array_reverse($tablesToJoin);
-        $tablesToJoin[] = $baseTable;
-
-        return $builder->aqb->merge(...$tablesToJoin);
-    }
 
     /**
      * Compile an update statement into SQL.
@@ -476,5 +394,4 @@ class Grammar extends FluentAqlGrammar
     {
         return $expression->getValue();
     }
-
 }
