+++ tests/Eloquent/RelationshipQueriesTest.php
@@ -8,6 +8,7 @@ use LaravelFreelancerNL\Aranguent\Eloquent\Model;
 use Mockery as m;
 use Tests\Setup\Database\Seeds\CharactersSeeder;
 use Tests\Setup\Database\Seeds\LocationsSeeder;
+use Tests\Setup\Database\Seeds\TaggablesSeeder;
 use Tests\Setup\Database\Seeds\TagsSeeder;
 use Tests\Setup\Models\Character;
 use Tests\TestCase;
@@ -20,8 +21,9 @@ class RelationshipQueriesTest extends TestCase
         Carbon::setTestNow(Carbon::now());
 
         Artisan::call('db:seed', ['--class' => CharactersSeeder::class]);
-        Artisan::call('db:seed', ['--class' => TagsSeeder::class]);
         Artisan::call('db:seed', ['--class' => LocationsSeeder::class]);
+        Artisan::call('db:seed', ['--class' => TagsSeeder::class]);
+        Artisan::call('db:seed', ['--class' => TaggablesSeeder::class]);
     }
 
     public function tearDown(): void
@@ -39,20 +41,23 @@ class RelationshipQueriesTest extends TestCase
         $this->assertEquals(3, count($characters));
     }
 
-    public function testDoesntHave()
-    {
-        $characters = Character::doesntHave('leads')->get();
-        $this->assertEquals(40, count($characters));
-    }
-
-    /*
-     * Requires override of getRelationExistenceCountQuery function on relation.php
-     * and thus of all child classes...
-     */
     public function testHasWithMinimumRelationCount()
     {
         $characters = Character::has('leads', '>=', 3)->get();
         $this->assertEquals(1, count($characters));
+    }
+
+    public function testHasMorph()
+    {
+        $characters = Character::has('tags')->get();
+
+        $this->assertEquals(2, count($characters));
+    }
+
+    public function testDoesntHave()
+    {
+        $characters = Character::doesntHave('leads')->get();
+        $this->assertEquals(40, count($characters));
     }
 
     public function testWithCount()
